feat: firebase auth edge login
