+++ libs/shared/ui/src/libs/crypto/decrypt.ts
@@ -1,19 +1,19 @@
-export async function decrypt(
+import crypto from 'crypto'
+
+export function decrypt(
   encrypt: [key: string, iv: string, encrypted: string]
-): Promise<string> {
+): string {
   const [key, iv, encrypted] = encrypt
-  const pwUtf8 = new TextEncoder().encode(key)
-  const ptUtf8 = new TextEncoder().encode(encrypted)
-  const pwHash = await crypto.subtle.digest('SHA-256', pwUtf8)
-  const alg = { name: 'AES-GCM', iv }
-  const decryptKey = await crypto.subtle.importKey('raw', pwHash, alg, false, [
-    'decrypt'
-  ])
-  const decipher = await crypto.subtle.decrypt(
+  const decipher = crypto.createDecipheriv(
     'aes-256-cbc',
-    decryptKey,
-    ptUtf8
+    key,
+    Buffer.from(iv, 'hex')
   )
 
-  return new TextDecoder().decode(decipher)
+  const decrpyted = Buffer.concat([
+    decipher.update(Buffer.from(encrypted, 'hex')),
+    decipher.final()
+  ])
+
+  return decrpyted.toString()
 }
